separate ObjectLinesHolder from ObjectLines
