+++ source/MRMesh/MRObjectLines.cpp
@@ -14,23 +14,23 @@ namespace MR
 
 MR_ADD_CLASS_FACTORY( ObjectLines )
 
-void ObjectLines::applyScale( float scaleFactor )
-{
-    if ( !polyline_ )
-        return;
-
-    auto& points = polyline_->points;
-
-    tbb::parallel_for( tbb::blocked_range<int>( 0, ( int )points.size() ),
-        [&] ( const tbb::blocked_range<int>& range )
-    {
-        for ( int i = range.begin(); i < range.end(); ++i )
-        {
-            points[VertId( i )] *= scaleFactor;
-        }
-    } );
-    setDirtyFlags( DIRTY_POSITION );
-}
+// void ObjectLines::applyScale( float scaleFactor )
+// {
+//     if ( !polyline_ )
+//         return;
+// 
+//     auto& points = polyline_->points;
+// 
+//     tbb::parallel_for( tbb::blocked_range<int>( 0, ( int )points.size() ),
+//         [&] ( const tbb::blocked_range<int>& range )
+//     {
+//         for ( int i = range.begin(); i < range.end(); ++i )
+//         {
+//             points[VertId( i )] *= scaleFactor;
+//         }
+//     } );
+//     setDirtyFlags( DIRTY_POSITION );
+// }
 
 std::shared_ptr<Object> ObjectLines::clone() const
 {
@@ -67,21 +67,21 @@ void ObjectLines::setDirtyFlags( uint32_t mask )
     }
 }
 
-void ObjectLines::setLineWidth( float width )
-{
-    if ( width == lineWidth_ )
-        return;
-    lineWidth_ = width;
-    needRedraw_ = true;
-}
-
-void ObjectLines::setPointSize( float size )
-{
-    if ( size == pointSize_ )
-        return;
-    pointSize_ = size;
-    needRedraw_ = true;
-}
+// void ObjectLines::setLineWidth( float width )
+// {
+//     if ( width == lineWidth_ )
+//         return;
+//     lineWidth_ = width;
+//     needRedraw_ = true;
+// }
+// 
+// void ObjectLines::setPointSize( float size )
+// {
+//     if ( size == pointSize_ )
+//         return;
+//     pointSize_ = size;
+//     needRedraw_ = true;
+// }
 
 void ObjectLines::swapBase_( Object& other )
 {
@@ -91,17 +91,17 @@ void ObjectLines::swapBase_( Object& other )
         assert( false );
 }
 
-Box3f ObjectLines::getWorldBox() const
-{
-    if ( !polyline_ )
-        return {};
-    const auto worldXf = this->worldXf();
-    if ( auto v = worldBox_.get( worldXf ) )
-        return *v;
-    const auto box = polyline_->computeBoundingBox( &worldXf );
-    worldBox_.set( worldXf, box );
-    return box;
-}
+// Box3f ObjectLines::getWorldBox() const
+// {
+//     if ( !polyline_ )
+//         return {};
+//     const auto worldXf = this->worldXf();
+//     if ( auto v = worldBox_.get( worldXf ) )
+//         return *v;
+//     const auto box = polyline_->computeBoundingBox( &worldXf );
+//     worldBox_.set( worldXf, box );
+//     return box;
+// }
 
 std::vector<std::string> ObjectLines::getInfoLines() const
 {
@@ -128,133 +128,133 @@ std::vector<std::string> ObjectLines::getInfoLines() const
     return res;
 }
 
-AllVisualizeProperties ObjectLines::getAllVisualizeProperties() const
+// AllVisualizeProperties ObjectLines::getAllVisualizeProperties() const
+// {
+//     AllVisualizeProperties res;
+//     res.resize( LinesVisualizePropertyType::LinesVisualizePropsCount );
+//     for ( int i = 0; i < res.size(); ++i )
+//         res[i] = getVisualizePropertyMask( unsigned( i ) );
+//     return res;
+// }
+// 
+// const ViewportMask& ObjectLines::getVisualizePropertyMask( unsigned type ) const
+// {
+//     switch ( LinesVisualizePropertyType::Type( type ) )
+//     {
+//     case MR::LinesVisualizePropertyType::Points:
+//         return showPoints_;
+//     case MR::LinesVisualizePropertyType::Smooth:
+//         return smoothConnections_;
+//     default:
+//         return VisualObject::getVisualizePropertyMask( type );
+//     }
+// }
+
+ObjectLines::ObjectLines( const ObjectLines& other ) :
+    ObjectLinesHolder( other )
 {
-    AllVisualizeProperties res;
-    res.resize( LinesVisualizePropertyType::LinesVisualizePropsCount );
-    for ( int i = 0; i < res.size(); ++i )
-        res[i] = getVisualizePropertyMask( unsigned( i ) );
-    return res;
 }
 
-const ViewportMask& ObjectLines::getVisualizePropertyMask( unsigned type ) const
-{
-    switch ( LinesVisualizePropertyType::Type( type ) )
-    {
-    case MR::LinesVisualizePropertyType::Points:
-        return showPoints_;
-    case MR::LinesVisualizePropertyType::Smooth:
-        return smoothConnections_;
-    default:
-        return VisualObject::getVisualizePropertyMask( type );
-    }
-}
-
-ObjectLines::ObjectLines()
-{
-    setDefaultColors_();
-}
-
-void ObjectLines::serializeFields_( Json::Value& root ) const
-{
-    VisualObject::serializeFields_(root);
-
-    root["ShowPoints"] = showPoints_.value();
-    root["SmoothConnections"] = smoothConnections_.value();
-
-    if ( !polyline_ )
-        return;
-    auto& polylineRoot = root["Polyline"];
-    auto& pointsRoot = polylineRoot["Points"];
-    auto& linesRoot = polylineRoot["Lines"];
-    for ( const auto& p : polyline_->points )
-    {
-        Json::Value val;
-        serializeToJson( p, val );
-        pointsRoot.append( val );
-    }
-    for ( UndirectedEdgeId ue{0}; ue < polyline_->topology.undirectedEdgeSize(); ++ue )
-    {
-        auto o = polyline_->topology.org( ue );
-        auto d = polyline_->topology.dest( ue );
-        if ( o && d )
-        {
-            linesRoot.append( int( o ) );
-            linesRoot.append( int( d ) );
-    }
-    }
-
-    // Type
-    root["Type"].append( ObjectLines::TypeName() ); // will be appended in derived calls
-}
-
-void ObjectLines::deserializeFields_( const Json::Value& root )
-{
-    VisualObject::deserializeFields_(root);
-
-    if ( root["ShowPoints"].isUInt() )
-        showPoints_ = ViewportMask{ root["ShowPoints"].asUInt() };
-    if ( root["SmoothConnections"].isUInt() )
-        smoothConnections_ = ViewportMask{ root["SmoothConnections"].asUInt() };
-
-    const auto& polylineRoot = root["Polyline"];
-    if ( !polylineRoot.isObject() )
-        return;
-
-    const auto& pointsRoot = polylineRoot["Points"];
-    const auto& linesRoot = polylineRoot["Lines"];
-
-    if ( !pointsRoot.isArray() || !linesRoot.isArray() )
-        return;
-
-    Polyline3 polyline;
-    polyline.points.resize( pointsRoot.size() );
-    for ( int i = 0; i < polyline.points.size(); ++i )
-        deserializeFromJson( pointsRoot[i], polyline.points.vec_[i] );
-
-    int maxVertId = -1;
-    for ( int i = 0; i < (int)linesRoot.size(); ++i )
-        maxVertId = std::max( maxVertId, linesRoot[i].asInt() );
-
-    polyline.topology.vertResize( maxVertId + 1 );
-    for ( int i = 0; i < (int)linesRoot.size(); i += 2 )
-        polyline.topology.makeEdge( VertId( linesRoot[i].asInt() ), VertId( linesRoot[i + 1].asInt() ) );
-
-    polyline_ = std::make_shared<Polyline3>( std::move( polyline ) );
-    setDirtyFlags( DIRTY_ALL );
-}
-
-Box3f ObjectLines::computeBoundingBox_() const
-{
-    if ( !polyline_ )
-        return {};
-    Box3f box;
-    for ( const auto& p : polyline_->points )
-        box.include( p );
-    return box;
-}
-
-Box3f ObjectLines::computeBoundingBoxXf_() const
-{
-    if( !polyline_ )
-        return {};
-    Box3f box;
-    auto wXf = worldXf();
-    for( const auto& p : polyline_->points )
-        box.include( wXf (p) );
-    return box;
-}
-
-void ObjectLines::setupRenderObject_() const
-{
-    if ( !renderObj_ )
-        renderObj_ = createRenderObject<ObjectLines>( *this );
-}
-
-void ObjectLines::setDefaultColors_()
-{
-    setFrontColor( SceneColors::get( SceneColors::SelectedObjectLines ) );
-    setFrontColor( SceneColors::get( SceneColors::UnselectedObjectLines ), false );
-}
+// void ObjectLines::serializeFields_( Json::Value& root ) const
+// {
+//     VisualObject::serializeFields_(root);
+// 
+//     root["ShowPoints"] = showPoints_.value();
+//     root["SmoothConnections"] = smoothConnections_.value();
+// 
+//     if ( !polyline_ )
+//         return;
+//     auto& polylineRoot = root["Polyline"];
+//     auto& pointsRoot = polylineRoot["Points"];
+//     auto& linesRoot = polylineRoot["Lines"];
+//     for ( const auto& p : polyline_->points )
+//     {
+//         Json::Value val;
+//         serializeToJson( p, val );
+//         pointsRoot.append( val );
+//     }
+//     for ( UndirectedEdgeId ue{0}; ue < polyline_->topology.undirectedEdgeSize(); ++ue )
+//     {
+//         auto o = polyline_->topology.org( ue );
+//         auto d = polyline_->topology.dest( ue );
+//         if ( o && d )
+//         {
+//             linesRoot.append( int( o ) );
+//             linesRoot.append( int( d ) );
+//     }
+//     }
+// 
+//     // Type
+//     root["Type"].append( ObjectLines::TypeName() ); // will be appended in derived calls
+// }
+// 
+// void ObjectLines::deserializeFields_( const Json::Value& root )
+// {
+//     VisualObject::deserializeFields_(root);
+// 
+//     if ( root["ShowPoints"].isUInt() )
+//         showPoints_ = ViewportMask{ root["ShowPoints"].asUInt() };
+//     if ( root["SmoothConnections"].isUInt() )
+//         smoothConnections_ = ViewportMask{ root["SmoothConnections"].asUInt() };
+// 
+//     const auto& polylineRoot = root["Polyline"];
+//     if ( !polylineRoot.isObject() )
+//         return;
+// 
+//     const auto& pointsRoot = polylineRoot["Points"];
+//     const auto& linesRoot = polylineRoot["Lines"];
+// 
+//     if ( !pointsRoot.isArray() || !linesRoot.isArray() )
+//         return;
+// 
+//     Polyline3 polyline;
+//     polyline.points.resize( pointsRoot.size() );
+//     for ( int i = 0; i < polyline.points.size(); ++i )
+//         deserializeFromJson( pointsRoot[i], polyline.points.vec_[i] );
+// 
+//     int maxVertId = -1;
+//     for ( int i = 0; i < (int)linesRoot.size(); ++i )
+//         maxVertId = std::max( maxVertId, linesRoot[i].asInt() );
+// 
+//     polyline.topology.vertResize( maxVertId + 1 );
+//     for ( int i = 0; i < (int)linesRoot.size(); i += 2 )
+//         polyline.topology.makeEdge( VertId( linesRoot[i].asInt() ), VertId( linesRoot[i + 1].asInt() ) );
+// 
+//     polyline_ = std::make_shared<Polyline3>( std::move( polyline ) );
+//     setDirtyFlags( DIRTY_ALL );
+// }
+// 
+// Box3f ObjectLines::computeBoundingBox_() const
+// {
+//     if ( !polyline_ )
+//         return {};
+//     Box3f box;
+//     for ( const auto& p : polyline_->points )
+//         box.include( p );
+//     return box;
+// }
+// 
+// Box3f ObjectLines::computeBoundingBoxXf_() const
+// {
+//     if( !polyline_ )
+//         return {};
+//     Box3f box;
+//     auto wXf = worldXf();
+//     for( const auto& p : polyline_->points )
+//         box.include( wXf (p) );
+//     return box;
+// }
+// 
+// void ObjectLines::setupRenderObject_() const
+// {
+//     if ( !renderObj_ )
+//         renderObj_ = createRenderObject<ObjectLines>( *this );
+// }
+// 
+// void ObjectLines::setDefaultColors_()
+// {
+//     setFrontColor( SceneColors::get( SceneColors::SelectedObjectLines ) );
+//     setFrontColor( SceneColors::get( SceneColors::UnselectedObjectLines ), false );
+// }
 
 }
